requirements.txt
# Core ADK and Gemini
google-genai>=0.2.0
google-generativeai>=0.8.0

# MCP Support
mcp>=0.1.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
structlog>=23.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
jsonschema>=4.19.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional deployment
uvicorn>=0.23.0
fastapi>=0.103.0

gradio>=3.0

tzlocal

